third_party/WebKit/LayoutTests/fast/forms/select/select-max-length-expected.txt

CONSOLE WARNING: line 10: Blocked to expand the option list to 2147483648 items.  The maximum list length is 2147483647.
CONSOLE WARNING: Blocked to expand the option list and set an option at index=2147483648.  The maximum list length is 2147483647.
CONSOLE WARNING: Blocked to expand the option list and set an option at index=2147483648.  The maximum list length is 2147483647.
This test that setting HTMLSelectElement.length is capped to 2,147,483,647, and you can't add additional Option elements too.

PASS sel.length is 0
Trying: - sel.length = 2147483648;
PASS sel.length is 0
Trying: - sel.add(new Option, 0);
PASS sel.length is 1
Trying: - sel.length = 0;
PASS sel.length is 0
Index setter:
PASS sel[2147483648] = new Option(); sel.length is 0
PASS sel.options[2147483648] = new Option(); sel.length is 0
PASS successfullyParsed is true

TEST COMPLETE
